platform/x86/amd/hsmp: Create separate ACPI, plat and common drivers
commit 7d3135d16356f1f0adda7e76d4a747f618263db4 upstream.

Separate the probes for HSMP ACPI and platform device drivers.

Provide a Kconfig option to choose between ACPI or the platform device
based driver. The common code which is the core part of the HSMP driver
maintained at hsmp.c is guarded by AMD_HSMP config and is selected by
these two driver configs. This will be built into separate hsmp_common.ko
module and acpi as hsmp_acpi and plat as amd_hsmp respectively.

Also add "|| COMPILE_TEST" clause in Kconfig to get build coverage for
HSMP.

[Backport changes]

Since there is no commit introducing the change from .remove_new to .remove,
we are removing .remove_new from drivers/platform/x86/amd/hsmp/hsmp.c.
As per upstream, the .remove was added in drivers/platform/x86/amd/hsmp/plat.c.

Link: https://lore.kernel.org/r/20241021111428.2676884-8-suma.hegde@amd.com
[ij: Fixed doc to use pre-formatted text for the ACPI dump.]
